backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2
email-validator==2.1.0.post1 